+++ common/src/main/java/com/nemonotfound/nemos/inventory/sorting/service/InventoryService.java
@@ -9,13 +9,11 @@ public class InventoryService {
     private final SortingService sortingService;
     private final MergingService mergeService;
     private final SplitQuickMoveService splitQuickMoveService;
-    private final ScrollTransferService scrollTransferService;
 
-    private InventoryService(MergingService mergeService, SortingService sortingService, SplitQuickMoveService splitQuickMoveService, ScrollTransferService scrollTransferService) {
+    private InventoryService(MergingService mergeService, SortingService sortingService, SplitQuickMoveService splitQuickMoveService) {
         this.mergeService = mergeService;
         this.sortingService = sortingService;
         this.splitQuickMoveService = splitQuickMoveService;
-        this.scrollTransferService = scrollTransferService;
     }
 
     public static InventoryService getInstance() {
@@ -23,8 +21,7 @@ public class InventoryService {
             INSTANCE = new InventoryService(
                     MergingService.getInstance(),
                     SortingService.getInstance(),
-                    SplitQuickMoveService.getInstance(),
-                    ScrollTransferService.getInstance()
+                    SplitQuickMoveService.getInstance()
             );
         }
 
@@ -42,9 +39,5 @@ public class InventoryService {
 
     public void handleSplitQuickMove(AbstractContainerMenu menu, int slot) {
         splitQuickMoveService.handleSplitQuickMove(menu, slot);
-    }
-
-    public boolean handleSingleItemScrollMove(AbstractContainerMenu menu, int slot, double scrollDelta) {
-        return scrollTransferService.handleSingleItemScrollMove(menu, slot, scrollDelta);
     }
 }